requirements.txt
Django==4.2.6
djangorestframework==3.14.0
PyMySQL==0.8.0
mysqlclient